requirements.txt
# Core dependencies
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.26
alembic==1.12.1
flask-security-too==5.3.0
Flask-Limiter==3.5.0

# Utilities
email-validator==2.1.0.post1
python-dotenv==1.0.0
Jinja2==3.1.2
itsdangerous==2.1.2
WTForms==3.0.1
chardet==5.2.0

# Optional database drivers
# Uncomment the one you need
# psycopg2-binary==2.9.9  # PostgreSQL (requires pg_config to be installed)
# pymysql==1.1.0  # MySQL
